requirements.txt
# Vector Databases
chromadb>=0.4.22
faiss-cpu>=1.7.4

# Embeddings
sentence-transformers>=2.3.0

# Local LLM Integration
ollama>=0.1.6
requests>=2.31.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0

# Chat Interface
gradio>=4.14.0

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.0

# Development (optional)
pytest>=7.4.0
black>=23.12.0
flake8>=7.0.0
